add auth on manager api's

+++ app/api/mixer/manager/user/deleteUser/route.js
@@ -1,9 +1,15 @@
 import { NextResponse } from 'next/server';
 import prisma from '@/prisma';
-
-export async function POST(request) {
+import { auth } from '@/auth';
+export const POST = auth(async function POST(req) {
+  if (!req.auth) {
+    return NextResponse.json({ success: false, error: 'Unauthorized' });
+  }
+  if (!req.auth.user.admin) {
+    return NextResponse.json({ success: false, error: 'Unauthorized' });
+  }
   try {
-    const { climberId } = await request.json();
+    const { climberId } = await req.json();
 
     await prisma.MixerClimber.delete({
       where: { id: climberId },
@@ -17,4 +23,4 @@ export async function POST(request) {
       message: 'error finding user in api',
     });
   }
-}
+});
